[CS-2708]: Use queryWrapper on claim revenue
